notification_service/requirements.txt
Jinja2==2.11.2
MarkupSafe==1.1.1
pika==1.1.0
# pkg-resources==0.0.0
python-dotenv==0.15.0
python-http-client==3.3.1
sendgrid==6.5.0
starkbank-ecdsa==1.1.0
